change MenuItem price type from Float to Int

+++ app/src/main/java/org/d3if2101/canteen/datamodels/CartItem.kt
@@ -4,7 +4,7 @@ data class CartItem(
     var itemID: String = "ITEM_ID",
     var imageUrl: String = "IMAGE_URL",
     var itemName: String = "ITEM_NAME",
-    var itemPrice: Float = 0.0F,
+    var itemPrice: Int = 0,
     var itemShortDesc: String = "ITEM_DESC",
     var itemStars: Float = 5.0F,
     var quantity: Int = 0
